Finished the authentacation routes
