+++ backend/controllers/session.js
@@ -1,11 +1,54 @@
 const jwt = require('jsonwebtoken');
 const userModels = require('../models/users.js');
+const crypto = require("crypto");
 
-const generateAuthToken = async (email,role) => {
+//checks if the email entered is valid
+const checkIfValid = (email) => {
+    const re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+    console.log("Is valid email:", re.test(String(email).toLowerCase()));
+    return re.test(String(email).toLowerCase());
+};
 
-    const accessTokenSecret = 'The President\'s Scholar stole the President\'s cake';
+
+const createNewUser = async (firstName, lastName, email, password) => {
+    if(firstName == null || lastName == null || email == null || password == null){
+        return {
+            error: "Incomplete input"
+        }
+    }
+    console.log('Raw password:', password);
+    //generates a random salt
+    const salt = "DB_SALT";
+    console.log('Password salt', salt);
+    // const hashedPassword = await bcrypt.hash(password, salt);
+    //hashes the password with the salt using sha256
+    const hashedPassword = crypto.createHash('sha256').update(salt + password).digest('hex');
+    console.log('Hashed password', hashedPassword);
+    //checks to see if email given is a valid email
+    const isValidEmail = checkIfValid(email);
+    if (!isValidEmail) {
+        return {
+            error: 'Invalid email'
+        }
+    }
+
+    //checks to see if the user already exists
+
+    const isUnique = await userModels.isUnique(email);
+    if (!isUnique) {
+        return {
+            error: 'User already exists'
+        }
+    }
+
+    const result = await userModels.createNewUser(firstName, lastName, email, hashedPassword,salt);
+    return result;
+}
+
+const generateAuthToken = async (email , role) => {
+
+    const accessTokenSecret = 'mysupercoolsecret';
     const users = await userModels.findUserByEmail(email);
-    //TODO claims when roles are implemented
     const accessToken = jwt.sign({...users[0], claims: [role]}, accessTokenSecret);
     return accessToken;
 
@@ -16,7 +59,7 @@ const authenticateUser = async (email, password) => {
     if(email == null || password == null) {
         return {
             error: 'Email or password is missing'
-        }
+        };
     }
     const users = await userModels.findUserByEmail(email);
     console.log('Results of users query', users);
@@ -25,7 +68,7 @@ const authenticateUser = async (email, password) => {
         console.error(`No users matched the email: ${email}`);
         return {
             error: 'Invalid credentials'
-        }
+        };
     }
     //if they do, checks to see if password is correct
     const authentication = await userModels.authenticateUser(email, password);
@@ -33,13 +76,15 @@ const authenticateUser = async (email, password) => {
         console.error(`Password for email: ${email} is incorrect`);
         return {
             error: 'Invalid credentials'
-        }
+        };
     }
     return {
         error: ''
-    }
+    };
 }
 
 module.exports = {
-    authenticateUser
+    authenticateUser,
+    generateAuthToken,
+    createNewUser
 }
